+++ crates/teloxide-core/CHANGELOG.md
@@ -127,6 +127,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `can_send_media_messages` field from `Restricted` ([#954][pr954])
 - Previously deprecated items ([#1013][pr1013])
   - `AutoSend` bot adaptor
+  - `ChatMemberKind::is_kicked` (use `is_banned` instead)
+  - `ChatMemberKind::is_creator` (use `is_owner` instead)
+  - `ChatMemberKind::{can_change_info, can_pin_messages, can_invite_users, can_manage_topics, can_send_polls, can_add_web_page_previews, can_send_other_messages, can_send_media_messages, can_send_messages}` (match on `ChatMemberKind` yourself)
+  - `ChatMemberStatus::is_present` (use `ChatMemberKind::is_present` instead)
 
 [pr954]: https://github.com/teloxide/teloxide/pull/954
 [pr1013]: https://github.com/teloxide/teloxide/pull/1013